Refatoração para separar em modulos
